endpoints adjusted for users

+++ routes/folders.js
@@ -10,7 +10,9 @@ const Note = require('../models/note');
 
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/folders', (req, res, next) => {
-  Folder.find()
+  const userId = req.user.id;
+
+  Folder.find({userId: userId})
     .sort('name')
     .then(results => {
       res.json(results);
@@ -23,6 +25,7 @@ router.get('/folders', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/folders/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -30,7 +33,7 @@ router.get('/folders/:id', (req, res, next) => {
     return next(err);
   }
 
-  Folder.findById(id)
+  Folder.findOne({_id: id, userId})
     .then(result => {
       if (result) {
         res.json(result);
@@ -46,8 +49,9 @@ router.get('/folders/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/folders', (req, res, next) => {
   const { name } = req.body;
+  const userId = req.user.id;
 
-  const newFolder = { name };
+  const newFolder = { name, userId };
 
   /***** Never trust users - validate input *****/
   if (!name) {
@@ -73,6 +77,7 @@ router.post('/folders', (req, res, next) => {
 router.put('/folders/:id', (req, res, next) => {
   const { id } = req.params;
   const { name } = req.body;
+  const userId = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!name) {
@@ -89,43 +94,71 @@ router.put('/folders/:id', (req, res, next) => {
 
   const updateFolder = { name };
 
-  Folder.findByIdAndUpdate(id, updateFolder, { new: true })
-    .then(result => {
-      if (result) {
-        res.json(result);
-      } else {
-        next();
+
+  Folder.findById(id)
+    .then((result) => {
+      if (String(result.userId) === String(userId)) {
+        Folder.findByIdAndUpdate(id, updateFolder, { new: true })
+        .then(result => {
+          if (result) {
+            res.json(result);
+          } else {
+            next();
+          }
+        })
+        .catch(err => {
+          if (err.code === 11000) {
+            err = new Error('The folder name already exists');
+            err.status = 400;
+          }
+          next(err);
+        });
+      }       
+      else {
+        const err = new Error('This id does not belong to this user.');
+        err.status = 400;
+        return next(err);
       }
     })
-    .catch(err => {
-      if (err.code === 11000) {
-        err = new Error('The folder name already exists');
-        err.status = 400;
-      }
+    .catch((err) => {
       next(err);
-    });
+    })
+
 });
 
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/folders/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
-  // Manual "cascading" delete to ensure integrity
-  const folderRemovePromise = Folder.findByIdAndRemove({ _id: id });  // NOTE **underscore** _id
-  // const noteRemovePromise = Note.deleteMany({ folderId: id });
+  Folder.findById(id)
+    .then((result) => {
+      if (String(result.userId) === String(userId)) {
+        // Manual "cascading" delete to ensure integrity
+        const folderRemovePromise = Folder.findByIdAndRemove({ _id: id });  // NOTE **underscore** _id
+        // const noteRemovePromise = Note.deleteMany({ folderId: id });
 
-  const noteRemovePromise = Note.updateMany(
-    { folderId: id },
-    { '$unset': { 'folderId': '' } }
-  );
+        const noteRemovePromise = Note.updateMany(
+          { folderId: id },
+          { '$unset': { 'folderId': '' } }
+        );
 
-  Promise.all([folderRemovePromise, noteRemovePromise])
-    .then(() => {
-      res.status(204).end();
+        Promise.all([folderRemovePromise, noteRemovePromise])
+          .then(() => {
+            res.status(204).end();
+          })
+          .catch(err => {
+            next(err);
+          });
+      }
+      else {
+        const err = new Error('This id does not belong to this user.');
+        err.status = 400;
+        return next(err);
+      }
     })
-    .catch(err => {
-      next(err);
-    });
+
+
 });
 
 module.exports = router;
